pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-transcription-tool"
version = "0.1.0"
description = "A unified speech-to-text script supporting local files and YouTube links using OpenAI Whisper API"
readme = "README.md"
license = "MIT"
authors = [
    {name = "LMFuture", email = "2835365572zty@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "boto3>=1.20.0",
    "botocore>=1.23.0",
    "halo>=0.0.31",
    "PyYAML>=6.0",
    "yt-dlp>=2023.1.6",
    "groq>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.910",
]

[project.scripts]
transcribe = "whisper_transcription.main:main"

[project.urls]
Homepage = "https://github.com/first-storm/whisper-script"
Repository = "https://github.com/first-storm/whisper-script"
Issues = "https://github.com/first-storm/whisper-script/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
